testing: move facts testing into correct unit dir
